Added start of template system and user system
Added javascript libraries and pdo wrapper

The site is ugly but working logically fine, for the new stuff. I understand how to protect some routes and can easily expand that to different roles.

I am automatically validating users right now as its a lot of extra work to do the gui screens for the other things, and to activate email