Now have a resource area in project to upload, view, and delete files
Changed nginx config to allow images and pdf in the repo resources folder for each project, and its on the same path as the project url, but with the resources.

Added new routes, and twig views

Added two new js libraries, swiper to view things, and dropzone to upload them

Added a helper method in the projects to commit easily, and a way to get all the files in the resources. Added support in the history code to view changes with files added or removed

Added way to copy long urls of the resources for ease of adding to bb code editor