Standard views and edits
Standard Views and Edits

Got plugin structure set up (it was hard to figure out)

has support for new js date editor and
new js function and php route support to save data

Created consistent way all standard can save and then call success or fail upward in a chain

Added meta edit and view

Expanded meta standard keys

Put shared tag logic in the tag helper, because new standard pages need it too

Added small method to utilities

Fixed bug in standard search if empty

Added more logic in the meta converter for dates, in case timestamp stored in long value

Regular tag attributes now respect the no serialization to json if a key matches a standard with that attribute. If so, value is not written

Improved js standard top functions


Added drop image to the meta edit, 

Now converts projects file and resource urls stored in the meta to be stubbed so that git cloning and moving will not break the links

Added pre and post-processing for any attribute, if they set up the options in the consts

Refactored project model by moving out a lot of file and resource methods to a new class, and a lot of other code now calls these the new way. This allows anyplace to get access to the project folder and url bases


added js library SimpleLightbox to show larger images from thumbnails

Tempus Dominus at https://getdatepicker.com/ for chosing date time in the meta box


Did git view and edit (added new field, dropped field in the standard def of git)

Did color edit

Added google fonts, and a new css to set font family. The css edit has a working font picker of about 20 google fonts, that works

Added Chroma JS library for color names in gui


Did the css edit and view


Added create and edit controls for standard to the tag edits. Hooked up the callback to refresh the tags after standard changed on screen

Put in some more info about it, like the list of ancestor tags that contribute to this

Added a delete icon, but no action yet

Made javascript to generate ancestor tags for a standard

Put ancestor tags in the edit boxes

Tag attributes , if empty, are not overwritten by parent attribute of same name now

Fixed too big selected parent in edit tag box

Removed css attribute as standard key for css (too confusing when trying to edit as it was partially hidden)


Added the standard views to the tag dialog, and redid the look and feel of the create/edit buttons. 

Added a way to delete standards from a tag (at least that level)

Fixed Standard code to delete a standard if all the attributes for a tag for it are gone

Added tag method to better remove attributes