0.5.1 git bit (#2)
Settings for projects using standards 

Now its easy to use the gui for any sort of standard data a project might make. The general user flow will be to store settings across several projects in the user project, and specific settings in the projects

There is a javascript function that can take a data setup and generate a dialog to show the views of a colletion of standards, and makes a fancy dropdown box to chose one. Also the user can clear the setting

Currently, the first setting being used is to replace the git settings in the project db table. These settings are now in the history page. The Project model now has some fancy constants to help standardize its settings. Probably will be more added later

Project model has some useful functions to get and set settings, as well as to get or create tags by name and attributes by name (useful as standard settings exists in system defined tags and attributes that might not exist yet)

Added to the git standard, a key for pushing automatically.
It can be checked in the git editor, and shows up as a magic wand icon in the git display near the git repo ssh link


Each setting exists in a tag that does not have to be part of the project, this is where the tag pointer comes in, so that a commonly used setting can be adjusted and then those changes are seen elsewhere. Also, permissions are always used when reading and writing settings. If you use a tag outside of the project, you must be able to read it to use it

Made new route to set project settings

Started to actually upgrade the version number a bit , now and then

Flow Attributes can now point to flow tags
Added support for this in the attributes and briefs

Made the User public profile more visually appealing

BUGFIX the pointee name and link was not showing up in the tag and attribute editors. Also have new tag pointer icon to show there

Improvement: now the pointee link is generated differently from the sql, and includes all four types. Also now have an icon showing the type, and all the links work

Bugfix, fixed up the links the attributes point to, as they are displayed in the GUI. Tag attributes now pull more general stuff from the db when created to help make links



Added new column for the tag pointer and did a bit of index refactoring in the text files

Improved the widely used project permission checking function to not require part of the slim framework to work, now its optional

Centralized another amazingly useful php function to handle ajax processing. It started life as part of the tag control, but now is shared in all the helpers in the base class. Did some renaming here too

Put the user selected meta into the regular profile page

Added a more general function to get user specific settings and data

Tags now have an easy way to see if they have a standard attribute or not, done by simply checking for the required keys of a standard type given