WIP #9 git reflow
Added new class FlowEntryYaml.php which has methods for getting valid entry folders and filtering with the db, and marking invalid folders with a file

see todo in the FlowEntryBase.php in the load function maybe ok to change there, but the new class needs to be added to use it.

In the FlowEntryFiles.php totally redid the previously simple function of get_entry_folder and added way to get calculated from the Interface too. New naming structure for entry directories. Able to move directory and upgrade name too

 So, with the class and change to the FlowEntryFiles.php, now possible to not use the entry-summary.yaml in the top directory except as reference, so it can be an array of the new class json. And when importing entries, can just find all valid folders, and if necessary rename them during the import. At the same time putting a flag in directories ignored

Fixed BUG from last time that prevented page loads

Fixed sometimes inaccurate commit messages for entry save

* Then implement the clone from git, the clone from local file, and add patch and create patch for last n commits

and that will be the end of 5.2